Refactor AST parsers
